Extract methods from agent to routing module
